+++ src/lib/components/chat/Messages/ResponseMessage.svelte
@@ -4,6 +4,7 @@
 
 	import { createEventDispatcher } from 'svelte';
 	import { onMount, tick, getContext } from 'svelte';
+	import { haptic } from '$lib/utils/hapticAction';
 	import type { Writable } from 'svelte/store';
 	import type { i18n as i18nType, t } from 'i18next';
 
@@ -974,6 +975,7 @@
 												on:click={() => {
 													editMessageHandler();
 												}}
+												use:haptic={'nudge'}
 											>
 												<svg
 													xmlns="http://www.w3.org/2000/svg"
@@ -1002,6 +1004,7 @@
 										on:click={() => {
 											copyToClipboard(message.content);
 										}}
+										use:haptic={'nudge'}
 									>
 										<svg
 											xmlns="http://www.w3.org/2000/svg"
@@ -1032,6 +1035,7 @@
 													toggleSpeakMessage();
 												}
 											}}
+											use:haptic={'nudge'}
 										>
 											{#if loadingSpeech}
 												<svg
@@ -1111,6 +1115,7 @@
 													generateImage(message);
 												}
 											}}
+											use:haptic={'nudge'}
 										>
 											{#if generatingImage}
 												<svg
@@ -1226,6 +1231,7 @@
 															?.scrollIntoView();
 													}, 0);
 												}}
+												use:haptic={'nudge'}
 											>
 												<svg
 													stroke="currentColor"
@@ -1262,6 +1268,7 @@
 															?.scrollIntoView();
 													}, 0);
 												}}
+												use:haptic={'nudge'}
 											>
 												<svg
 													stroke="currentColor"
@@ -1292,6 +1299,7 @@
 												on:click={() => {
 													continueResponse();
 												}}
+												use:haptic={'success'}
 											>
 												<svg
 													xmlns="http://www.w3.org/2000/svg"
@@ -1338,6 +1346,7 @@
 													});
 												});
 											}}
+											use:haptic={'success'}
 										>
 											<svg
 												xmlns="http://www.w3.org/2000/svg"